Add API for adding a new user to the users.json file.
discord/interpreter.js has been edited to include a new API that allows
a user to add members to the users file. The command will check a list
of blacklisted names to ensure that a user cannot add any names that
shouldn't be added, for example bots. The command can use both the
@mention and the normal nickname. The addUser function in auth.js has
been exported to allow this new command to call it.