Update insert_user command
The insert_user command now will assign the member role to the mentioned
user if they do not already have it. Another npm package promise-retry
has been added to allow the promises to retry a certain number of times.
This could be useful if a failure occured due to network errors. Also:
- Updated Bot's messages to use the roles mentioned in the configuration
  file.