Add new Hakase responses
New file hakase_responses.js will be used to store responses
specific to Hakase. A few improvements also made to the welcome
command. Now it will check that whether the message is made in a
direct message channel or in a server. Depending on the channel
type, Hakase will say different things.